+++ src/LanguageServer.Engine/Logging/LoggerConfigurationLanguageServerExtensions.cs
@@ -1,8 +1,11 @@
 using Serilog;
 using Serilog.Configuration;
+using System;
 
 namespace MSBuildProjectTools.LanguageServer.Logging
 {
+    using Handlers;
+
     /// <summary>
     ///     Extension methods for configuring Serilog to log to a language server.
     /// </summary>
@@ -17,13 +20,25 @@ namespace MSBuildProjectTools.LanguageServer.Logging
         /// <param name="languageServer">
         ///     The language server to which events will be logged.
         /// </param>
+        /// <param name="configuration">
+        ///     The language server configuration handler.
+        /// </param>
         /// <returns>
         ///     The logger configuration.
         /// </returns>
-        public static LoggerConfiguration LanguageServer(this LoggerSinkConfiguration loggerSinkConfiguration, Lsp.LanguageServer languageServer)
+        public static LoggerConfiguration LanguageServer(this LoggerSinkConfiguration loggerSinkConfiguration, Lsp.LanguageServer languageServer, ConfigurationHandler configuration)
         {
+            if (loggerSinkConfiguration == null)
+                throw new ArgumentNullException(nameof(loggerSinkConfiguration));
+            
+            if (languageServer == null)
+                throw new ArgumentNullException(nameof(languageServer));
+            
+            if (configuration == null)
+                throw new ArgumentNullException(nameof(configuration));
+
             return loggerSinkConfiguration.Sink(
-                new LanguageServerLoggingSink(languageServer)
+                new LanguageServerLoggingSink(languageServer, configuration)
             );
         }
     }